Storing new submissions to mongodb
Updates collector will also generate store events, signyfing that
a new problem should be added for this user. For leetcode we are
only storing the total number of submissions, so it is a matter
of updating a counter once, which we can do once per scheduler execution per user.
For vjudge we need to store every solved problem to figure out the differences,
so we need a store event for each new solved problem.

Added models for leetcode and vjudge save events and a repository for submissions.

The events approach here allows for easy de-coupling of the updates collector service
and the main backend.